separate tests into main and functionality into parseMail.c
